added some code woooo
I started working on how to create a GUI (it took me a while to get the hang of). I also did some very minimal work on preprocessing. I created a new preprocessing.py file to put my code in, because I was not sure if we wanted to use the ipynb for coding or not.